pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "azure-cosmos>=4.14.1",
    "fastapi[standard]>=0.121.1",
    "google-api-python-client>=2.187.0",
    "google-api-python-client-stubs>=1.31.0",
    "python-dotenv>=1.2.1",
]
